pyproject.toml
[project]
name = "hh"
version = "0.1.0"
description = ""
authors = [
    {name = "Michael Keßler"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "django-tinymce (>=4.1.0,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
